requirements.txt
# Core dependencies
flask>=2.0.0
Flask-Limiter>=3.5.0
pydantic>=2.4.0
python-dotenv>=0.19.0
werkzeug>=2.0.0
gunicorn>=20.1.0

# API and external services
plaid-python>=9.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
selenium>=4.0.0

# File processing
PyPDF2>=3.0.0

# Data validation and security
email-validator>=2.0.0
bleach>=6.0.0

# OAuth and Google APIs
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0